Add files via upload
This project is an exploratory data analysis of a job listings dataset. It begins with Data Exploration (Phase 1), where the dataset's structure, data types, and basic statistics are examined. The dataset contains 79,147 entries across 12 columns, with some columns, like 'rating' and 'reviews,' having a significant number of missing values.

Next, the Data Cleaning (Phase 2) is performed, which involves several steps. Unnecessary columns such as 'posted_on,' 'job_link,' and 'company_link' are removed. Missing values are handled, duplicates are removed, and data types are standardized. For instance, the 'experience' and 'salary' columns are split into 'min_experience,' 'max_experience,' 'min_salary,' and 'max_salary' to convert string data into a numerical format for better analysis. The 'reviews' column, initially a string, is also converted to an integer.

The final part of the project is the Data Analysis and Visualization (Phase 3 and 4). This phase aims to answer key questions about the job market. Visualizations like bar charts and pie charts are used to show the total number of jobs and companies, identify the top companies with the most job openings and reviews, and highlight the most common skills required for different job roles, such as 'Data Analyst'.